backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
APScheduler==3.10.4
groq==0.11.0